fix(metrology): implement master PCR PID locking and industrial-grade bitrate settlement
- Implemented Master PCR PID locking in tsa_engine_pcr.c to prevent STC collisions in MPTS streams.
- Aligned physical bitrate calculation with Metrology Standards:
  - Use CLOCK_MONOTONIC in LIVE mode to avoid processing burst bias.
  - Enforce a 500ms minimum sampling window for stability.
  - Apply 20% instantaneous / 80% historical EMA smoothing.
- Fixed aggregate business bitrate (pcr_bitrate_bps) double-counting and SPTS/MPTS collision issues.
- Optimized TsPacer (tsp) stability:
  - Reduced default RING_BUFFER_SIZE to 16k to limit burst capacity.
  - Improved loop/lag synchronization to prevent 2Gbps line-speed surges.
- Updated verify_realtime_metrology.sh tolerance to 9.0-11.0 Mbps for real-time E2E stability.
- Enhanced unit tests (test_tsp_pcr_pacing, test_unit_metrology_repro) for better coverage of MPTS and pacing scenarios.

+++ tests/test_unit_pcr_drift.c
@@ -1,7 +1,7 @@
 #include <assert.h>
+#include <math.h>
 #include <stdio.h>
 #include <string.h>
-#include <math.h>
 
 #include "tsa.h"
 #include "tsa_internal.h"
@@ -77,8 +77,8 @@ int main() {
     /* 🚨 VERIFICATION: Should be close to 100 PPM after enough samples for EMA to converge */
 
     for (int loop = 2; loop <= 100; loop++) {
-        uint64_t wall_ns = 1000000000ULL + (loop * 1000000000ULL);
-        uint64_t pcr = loop * 27002700;
+        uint64_t wall_ns = 1000000000ULL + ((uint64_t)loop * 1000000000ULL);
+        uint64_t pcr = (uint64_t)loop * 27002700ULL;
         make_pcr_packet(pkt, 0x100, pcr, loop % 16);
         tsa_process_packet(h, pkt, wall_ns);
     }
